--- app/src/main/java/nextstep/signup/model/inputresult/InputValidity.kt
@@ -1,3 +1,0 @@
-package nextstep.signup.model.inputresult
-
-interface InputValidity
