+++ app/src/main/java/nextstep/signup/model/inputvalidity/EmailInputValidity.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.model.inputresult
+package nextstep.signup.model.inputvalidity
 
 enum class EmailInputValidity : InputValidity {
     INVALID_FORMAT,
