+++ app/src/main/java/nextstep/signup/model/inputvalidity/InputValidity.kt
@@ -1,0 +1,3 @@
+package nextstep.signup.model.inputvalidity
+
+interface InputValidity
